packages/agentbench/pyproject.toml
[project]
name = "agentbench"
version = "0.1.0"
description = "Offline evaluation harness for AI coding agents"
requires-python = ">=3.11"
dependencies = [
    "shared",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.24.0",
    "pyyaml>=6.0.0",
    "python-ulid>=2.0.0",
    "path>=17.1.1",
    "filelock>=3.20.1",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "pytest-asyncio>=0.21.0", "ruff>=0.1.0"]

[project.scripts]
ab = "agentbench.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
shared = { workspace = true }
